renamed vec_math.h back to math.h
