+++ samples/collection/shape_cast.cpp
@@ -4,7 +4,7 @@
 #include "sample.h"
 #include "box2d/constants.h"
 #include "box2d/distance.h"
-#include "box2d/vec_math.h"
+#include "box2d/math.h"
 
 class ShapeCast : public Sample
 {
