+++ src/distance.c
@@ -4,7 +4,7 @@
 #include "box2d/constants.h"
 #include "box2d/distance.h"
 #include "box2d/timer.h"
-#include "box2d/vec_math.h"
+#include "box2d/math.h"
 
 #include <assert.h>
 #include <float.h>
