+++ test/math_test.c
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Erin Catto
 // SPDX-License-Identifier: MIT
 
-#include "box2d/vec_math.h"
+#include "box2d/math.h"
 #include "test_macros.h"
 
 int MathTest()
