+++ test/shape_test.c
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "box2d/shapes.h"
-#include "box2d/vec_math.h"
+#include "box2d/math.h"
 #include "test_macros.h"
 
 #include <float.h>
